refactor(ddd): complete remaining service implementation migration
