+++ koduck-backend/src/main/java/com/koduck/identity/application/CredentialServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.identity.application;
 
 import com.koduck.common.constants.HttpHeaderConstants;
 import com.koduck.dto.credential.*;
