+++ koduck-backend/src/main/java/com/koduck/identity/application/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.identity.application;
 
 import com.koduck.dto.common.PageResponse;
 import com.koduck.dto.user.ChangePasswordRequest;
