+++ koduck-backend/src/main/java/com/koduck/market/application/KlineSyncServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.market.application;
 
 import com.koduck.common.constants.MarketConstants;
 import com.koduck.config.properties.DataServiceProperties;
