+++ koduck-backend/src/main/java/com/koduck/market/application/MarketFlowServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.market.application;
 import com.koduck.dto.market.DailyNetFlowDto;
 import com.koduck.mapper.MarketDataMapper;
 import com.koduck.repository.MarketDailyNetFlowRepository;
