+++ koduck-backend/src/main/java/com/koduck/market/application/SyntheticTickServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.market.application;
 
 import com.koduck.dto.market.TickDto;
 import com.koduck.entity.StockRealtime;
