+++ koduck-backend/src/main/java/com/koduck/shared/application/EmailServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.shared.application;
 
 import com.koduck.config.properties.MailProperties;
 import com.koduck.service.EmailService;
