+++ koduck-backend/src/main/java/com/koduck/shared/application/MemoryServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.shared.application;
 
 import com.koduck.entity.MemoryChatMessage;
 import com.koduck.entity.MemoryChatSession;
